Prevent `rill.yaml` renames and deletes
