+++ web-common/src/features/editor/FileWorkspaceHeader.svelte
@@ -4,11 +4,14 @@
   import { splitFolderAndName } from "@rilldata/web-common/features/sources/extract-file-name";
   import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
   import { WorkspaceHeader } from "../../layout/workspace";
+  import { PROTECTED_FILES } from "../file-explorer/protected-paths";
 
   export let filePath: string;
 
   let fileName: string;
+
   $: [, fileName] = splitFolderAndName(filePath);
+  $: isProtectedFile = PROTECTED_FILES.includes(filePath);
 
   const onChangeCallback = async (
     e: Event & {
@@ -29,4 +32,5 @@
   on:change={onChangeCallback}
   showInspectorToggle={false}
   titleInput={fileName}
+  editable={!isProtectedFile}
 />
