+++ web-common/src/features/file-explorer/protected-paths.ts
@@ -1,1 +1,2 @@
 export const PROTECTED_DIRECTORIES = ["/tmp"];
+export const PROTECTED_FILES = ["/rill.yaml"];
